[refactor] move file for refactoring
